requirements.txt
attrs==21.2.0
iniconfig==1.1.1
packaging==21.2
pluggy==1.0.0
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
pytest-mock==3.6.1
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
